core|tests: Relocate tests to sibling directory of tested files
