+++ packages/core/vitest.config.ts
@@ -1,9 +1,1 @@
-import { mergeConfig } from 'vitest/config';
-
-import baseConfig from '../../config/vitest.config.js';
-
-export default mergeConfig(baseConfig, {
-  test: {
-    passWithNoTests: true, // core has no tests; remove when it acquires source
-  },
-});
+export { default } from '../../config/vitest.config.ts';
